requirements.txt
opencv-python
numpy
matplotlib
seaborn
Pillow
tqdm
tensorboard
transformers
diffusers
accelerate
ipykernel
jupyter